requirements.txt
ccxt==4.1.13
python-dotenv==1.0.0
ratelimit==2.2.1
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
